requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
openai-whisper>=20231117
kokoro-onnx>=0.3.0
soundfile>=0.12.1
numpy>=1.24.0
python-multipart>=0.0.9
